# marketing-agency/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Marketing Agency API",
    "version": "0.1.0"
  },
  "paths": {
    "/research/brand/{brand_id}": {
      "post": {
        "summary": "Research From Brand Profile",
        "description": "Run research analysis using stored brand and product data",
        "operationId": "research_from_brand_profile_research_brand__brand_id__post",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "name": "is_new_brand",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": false,
              "title": "Is New Brand"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "user_1",
              "title": "User Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "session_001",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "summary": "Query Agent",
        "description": "Process a query using the selected marketing agent",
        "operationId": "query_agent_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research": {
      "post": {
        "summary": "Research Brand",
        "description": "Run brand research using the research and formatter agents",
        "operationId": "research_brand_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/{research_id}": {
      "get": {
        "summary": "Get Research",
        "description": "Retrieve a specific research by ID",
        "operationId": "get_research_research__research_id__get",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/brand/{brand_name}": {
      "get": {
        "summary": "Get Brand Research",
        "description": "Retrieve all research for a specific brand",
        "operationId": "get_brand_research_research_brand__brand_name__get",
        "parameters": [
          {
            "name": "brand_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Name"
            }
          },
          {
            "name": "product_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by product name",
              "title": "Product Name"
            },
            "description": "Filter by product name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResearchResponse"
                  },
                  "title": "Response Get Brand Research Research Brand  Brand Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/brand/{brand_name}/product/{product_name}": {
      "get": {
        "summary": "Get Product Research Endpoint",
        "description": "Retrieve research for a specific product of a brand",
        "operationId": "get_product_research_endpoint_research_brand__brand_name__product__product_name__get",
        "parameters": [
          {
            "name": "brand_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Name"
            }
          },
          {
            "name": "product_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Product Research Endpoint Research Brand  Brand Name  Product  Product Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/brand": {
      "post": {
        "summary": "Create Brand Profile",
        "description": "Create a brand profile with name and description",
        "operationId": "create_brand_profile_brand_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/brand/{brand_id}": {
      "get": {
        "summary": "Get Brand Profile",
        "description": "Get a brand profile by ID",
        "operationId": "get_brand_profile_brand__brand_id__get",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResearchResponse": {
        "properties": {
          "research_id": {
            "type": "string",
            "title": "Research Id"
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name"
          },
          "is_new_brand": {
            "type": "boolean",
            "title": "Is New Brand"
          },
          "research_data": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Research Data"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "research_id",
          "brand_name",
          "product_name",
          "is_new_brand",
          "research_data",
          "timestamp"
        ],
        "title": "ResearchResponse"
      },
...
    }
  }
}